Temporal module: restructuration according to the MVVM design pattern
